# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PairEval API
  version: 0.1.0-M1
  license:
    name: MIT
    identifier: MIT
  description: |
    API ของระบบประเมินผลงานนักศึกษาแบบ pairwise comparison

    **ขอบเขตของเอกสารนี้คือ M1 Walking Skeleton** — ทุก endpoint ที่นี่สืบกลับไปหา
    user story ใน `docs/backlog.md` ได้ และทุก story ใน Sprint 1 มี endpoint รองรับ

    หลักที่ยึดทั้งไฟล์
    - ทุก endpoint ยกเว้น `/api/health` ต้องมี bearer token
    - ตรวจสิทธิ์ที่ server เสมอ (FR-AUTHZ-01) — ห้ามพึ่ง UI ในการซ่อน
    - ขอ resource ของ classroom ที่ไม่ได้สังกัด ตอบ **404 ไม่ใช่ 403**
      เพื่อไม่ให้รู้ว่า resource นั้นมีอยู่จริง (FR-AUTHZ-02)
    - error ใช้รูปแบบเดียวกันทั้งระบบ และ `code` คงที่สำหรับเครื่องอ่าน (FR-API-03)
    - timestamp ทุกตัวเป็น UTC (DR-05)
servers:
  - url: https://paireval-api.onrender.com
    description: Staging
  - url: http://localhost:8000
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/classrooms/{classroomId}/assignments:
    get:
      tags: [classrooms]
      summary: รายการงานประเมินในห้องนี้
      description: |
        อ้าง US-07 — **ไม่ได้ระบุไว้ตรง ๆ ใน PRD §12** เป็นช่องว่างเดียวกับที่ US-01/US-12
        เคยเจอตอนไล่ตาราง traceability: ไม่มีทางให้นักศึกษารู้ id ของ assignment
        เพื่อเรียก `/my-evaluations` ต่อได้เลยถ้าไม่มี endpoint นี้

        นักศึกษาไม่เห็นงานที่ยังเป็น `DRAFT` เพราะยังไม่ควรรู้ด้วยซ้ำว่ามีงานนี้อยู่
        ผู้สอน (มี `MANAGE_ASSIGNMENT`) เห็นทุกสถานะเพื่อกลับมาแก้งานที่ทำค้างไว้ได้
      operationId: listClassroomAssignments
      parameters:
        - $ref: '#/components/parameters/ClassroomId'
      responses:
        '200':
          description: รายการงานประเมิน
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssignmentSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/assignments:
    post:
      tags: [assignments]
      summary: สร้างงานประเมิน
      description: |
        อ้าง US-04 — งานที่สร้างใหม่มีสถานะ `DRAFT` เสมอ
        ผลรวมน้ำหนักของ criteria แต่ละ side จะถูกตรวจตอน publish ไม่ใช่ตอนสร้าง
      operationId: createAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
      responses:
        '201':
          description: สร้างสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/assignments/{assignmentId}:
    get:
      tags: [assignments]
      summary: ดูงานประเมินหนึ่งงานพร้อมเกณฑ์
      description: 'อ้าง US-04'
      operationId: getAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: งานประเมิน
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/assignments/{assignmentId}/criteria:
    put:
      tags: [assignments]
      summary: แทนที่เกณฑ์ทั้งชุด
      description: |
        อ้าง US-04 — ทำได้เฉพาะตอนสถานะยังเป็น `DRAFT`
        งานที่ประกาศไปแล้วอาจมีคนเริ่มประเมินตามเกณฑ์ชุดเดิม การเปลี่ยนน้ำหนักกลางคัน
        ทำให้คะแนนจากคำตอบเก่ากับใหม่เทียบกันไม่ได้ จึงตอบ `409` แทนที่จะยอมให้แก้

        ส่งทั้งชุดไม่ใช่ทีละข้อ เพราะหน้าจอที่แก้เกณฑ์ในตารางเดียวมักยังไม่มี id
        ของแถวที่เพิ่งเพิ่ม การส่งทั้งชุดจึงตรงกับวิธีที่ UI ใช้จริงกว่า
      operationId: replaceCriteria
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [criteria]
              properties:
                criteria:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/CriterionCreate'
      responses:
        '200':
          description: เกณฑ์ชุดใหม่
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/assignments/{assignmentId}/feasibility:
    get:
      tags: [assignments]
      summary: ตรวจความเป็นไปได้ของการจัดคู่ก่อน publish
      description: |
        อ้าง US-05 — ถ้า `target_coverage` ที่ตั้งไว้เป็นไปไม่ได้
        ระบบคืนค่าสูงสุดที่ทำได้จริงพร้อมเหตุผลเป็นตัวเลข (FR-PAIR-05)
      operationId: getFeasibility
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: |
            ผลการคำนวณความเป็นไปได้ — คืน **ทั้งสองฝั่ง** เมื่อเปิดการประเมินรายบุคคลไว้
            เพราะข้อจำกัดของสองฝั่งคนละเรื่องกัน: ฝั่งกลุ่มถูกบีบด้วยจำนวนกลุ่มและ workload
            ส่วนฝั่งบุคคลถูกบีบด้วยขนาดกลุ่มล้วน ๆ (D4)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feasibility'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/assignments/{assignmentId}:publish:
    post:
      tags: [assignments]
      summary: เผยแพร่งานและสร้างคู่ประเมินทั้งหมด
      description: |
        อ้าง US-06 — สร้าง pair ทั้งหมดเก็บลง database ตอนนี้ ไม่สุ่มตอน runtime (FR-PAIR-01)
        ให้ `pairing_seed` เดิมแล้วผลต้องเหมือนเดิมทุกครั้ง (FR-PAIR-09)
      operationId: publishAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: เผยแพร่สำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'สถานะไม่ใช่ DRAFT จึง publish ไม่ได้ (`code: INVALID_STATE`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'ผลรวมน้ำหนัก criteria ไม่เท่ากับ 100 (`code: WEIGHT_SUM_INVALID`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/assignments/{assignmentId}/my-evaluations:
    get:
      tags: [evaluation]
      summary: คู่ที่ฉันต้องประเมินและความคืบหน้า
      description: |
        อ้าง US-07 — คืนเฉพาะคู่ของผู้เรียกเท่านั้น เปิดให้สมาชิกทุก role ของห้องเรียน
        ไม่ใช่แค่ผู้สอน เพราะนักศึกษาคือผู้ใช้หลักของ endpoint นี้

        **ยังไม่เปิด (`DRAFT`) ไม่ใช่ error** — ตอบ `200` เสมอพร้อม `opened: false`
        และ `message` อธิบายเหตุผล เพื่อให้หน้าเว็บ render ข้อความปกติได้ทันที
        ไม่ต้องแยก branch จัดการ error สำหรับสถานะที่คาดไว้อยู่แล้ว
      operationId: getMyEvaluations
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
        - name: side
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Side'
      responses:
        '200':
          description: 'รายการคู่ที่ต้องประเมิน (หรือ opened=false ถ้ายังไม่เปิด)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyEvaluations'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/assignments/{assignmentId}/submissions:
    post:
      tags: [evaluation]
      summary: ส่งคำตอบทั้ง side
      description: |
        อ้าง US-09 — เปลี่ยนทุก comparison ของ side นั้นที่**เคยตอบไว้แล้ว**เป็น `SUBMITTED`
        ต้องส่ง header `Idempotency-Key` เพื่อกันการกดซ้ำจากหลายแท็บ (FR-API-02) —
        เรียกซ้ำด้วย key เดิมได้ response เดิมทุกครั้ง (`submittedCount`/`submittedAt` ไม่ขยับ)

        **ตอบไม่ครบไม่ใช่ error** — เคยเขียนไว้ว่า 422 แต่ขัดกับ FR-EVAL-05 ที่บังคับให้ปุ่มส่ง
        เปิดใช้ได้ตลอดแม้ตอบไม่ครบ (ดูบันทึกการตัดสินใจใน `docs/backlog.md` ที่ US-09)
        endpoint นี้จึง submit เท่าที่มีเสมอ คู่ที่ไม่เคยตอบเลยไม่ถูกแตะ ไม่ใช่ปฏิเสธทั้งคำขอ
        ฝั่งหน้าเว็บเป็นคนแสดง "ยังเหลือ N คู่ ยืนยันจะส่งไหม" ก่อนเรียก endpoint นี้เอง
      operationId: submitEvaluations
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
        - name: Idempotency-Key
          in: header
          required: true
          description: ค่าไม่ซ้ำที่ client สร้างขึ้นต่อการกดส่ง 1 ครั้ง
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [side]
              properties:
                side:
                  $ref: '#/components/schemas/Side'
      responses:
        '200':
          description: ส่งสำเร็จ (นับเฉพาะคู่ที่เคยตอบไว้ — อาจน้อยกว่าจำนวนคู่ทั้งหมด)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'เลย deadline แล้ว (`code: DEADLINE_PASSED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/assignments/{assignmentId}:recompute:
    post:
      tags: [scoring]
      summary: คำนวณคะแนนชั่วคราวใหม่
      description: |
        อ้าง US-13, US-16 — คำนวณด้วย `is_final=false` ทับ interim เดิมเสมอ เรียกได้ตลอด
        หลัง publish ไม่ต้องรอ deadline (FR-SCORE-06) ใช้ดูความคืบหน้าระหว่างทาง
        ไม่เปลี่ยนสถานะ assignment และไม่ถูก audit เพราะยังไม่ใช่คะแนนที่เชื่อถือได้
      operationId: recomputeScores
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: คำนวณเสร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecomputeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `MANAGE_ASSIGNMENT` (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/assignments/{assignmentId}:finalize:
    post:
      tags: [scoring]
      summary: ตัดสินและประกาศคะแนน
      description: |
        อ้าง US-13 — เฉพาะ OWNER (capability `FINALIZE_SCORES`) เรียกได้ ต้องเลย
        `groupDeadlineUtc` ไปแล้วและ assignment อยู่ในสถานะ `PUBLISHED`/`OPEN`/`CLOSED`
        คำนวณ `is_final=true` เป็น append-only batch ใหม่ และเก็บ `score_snapshot` ที่มี
        input/output ครบชุด — ตรวจย้อนหลังได้แม้ reopen/finalize ใหม่หรือสูตรเปลี่ยนแล้ว

        ถ้ามี item ที่ผู้ประเมินยังไม่ถึงเกณฑ์ขั้นต่ำ (`LOW_CONFIDENCE`) endpoint นี้จะปฏิเสธ
        ด้วย 422 จนกว่าจะส่ง `confirmLowConfidence: true` มายืนยัน — ระบบไม่ finalize ให้อัตโนมัติ
      operationId: finalizeAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeAssignment'
      responses:
        '200':
          description: ประกาศผลแล้ว
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `FINALIZE_SCORES` — มีเฉพาะ OWNER (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'สถานะไม่ใช่ PUBLISHED/OPEN/CLOSED หรือยังไม่ถึง deadline (`code: ALREADY_EXISTS`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'มี item ติด LOW_CONFIDENCE และยังไม่ได้ส่ง confirmLowConfidence (`code: VALIDATION_FAILED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/assignments/{assignmentId}:reopen:
    post:
      tags: [scoring]
      summary: เปิดคะแนนที่ finalize แล้วกลับมาแก้
      description: |
        อ้าง US-13 — เฉพาะ OWNER เรียกได้ ต้องเป็นงานที่สถานะ `FINALIZED` เท่านั้น
        สถานะหลัง reopen กลับไปเป็น `CLOSED` (ไม่ใช่ `PUBLISHED`) ต้องเรียก `:finalize` ใหม่
        อีกครั้งเพื่อประกาศผล ต้องระบุเหตุผลเสมอเพราะกระทบความเชื่อมั่นของนักศึกษาที่เห็นคะแนน
        ไปแล้ว (`AuditAction.SCORES_REOPENED` อยู่ใน `ACTIONS_REQUIRING_REASON`)
      operationId: reopenAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenAssignment'
      responses:
        '200':
          description: เปิดกลับมาแก้แล้ว
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [CLOSED]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `FINALIZE_SCORES` (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'ยังไม่ได้ finalize (`code: ALREADY_EXISTS`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'ไม่ได้ระบุเหตุผล (`code: VALIDATION_FAILED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token ที่ได้หลัง login ผ่าน Google OIDC (US-01)
  parameters:
    ClassroomId:
      name: classroomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: 'ไม่ได้แนบ token หรือ token หมดอายุ (`code: UNAUTHENTICATED`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: |
        ไม่พบ resource — **รวมถึงกรณีที่ resource มีอยู่จริงแต่อยู่คนละ classroom**
        ตั้งใจตอบ 404 แทน 403 เพื่อไม่ให้รู้ว่ามีอยู่จริง (FR-AUTHZ-02)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    Assignment:
      type: object
      required: [id, classroomId, name, status, criteria]
      properties:
        id:
          type: string
          format: uuid
        classroomId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OPEN, CLOSED, FINALIZED, ARCHIVED]
        groupDeadlineUtc:
          type: string
          format: date-time
        individualDeadlineUtc:
          type: [string, 'null']
          format: date-time
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/Criterion'
    AssignmentSummary:
      type: object
      description: รูปย่อของ Assignment สำหรับหน้ารายการ — ไม่มี criteria เพราะหน้ารายการไม่ใช้
      required: [id, name, status, groupDeadlineUtc]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OPEN, CLOSED, FINALIZED, ARCHIVED]
        groupDeadlineUtc:
          type: string
          format: date-time
        individualDeadlineUtc:
          type: [string, 'null']
          format: date-time
...